Turn automuxerbin element into an auto adaptative one
Current implementation only provides a source pad when both audio and video
media are requested to the bin, no more requested pads are taken into account.
This version provides an output source pad as soon as a new pad is requested,
and it reconfigures itself when a new requested pad callback is managed, no
matter what kind of media it is, audio, video or even subtittles are supported.

Change-Id: I80220ca5c5661b14b906ed1922582c84f9116bc2