Change automuxerbin sink pad names
Many muxers use audio_%u, video_%u and so on for their sink pads names without
using the extra tag "sink". It would be more natural from the point of view of
a gstreamer developer to change it to be along the lines of the ones used by
other muxers.

Change-Id: If830ce0648aea7a937376488d2f9698e04a09632