Send an EOS event to the muxer that is going to be replaced
This patch forces the muxer to drain its data before being removed.
This patch also fixes the bug concerning the removal of the muxer
when it has not reached the NULL state yet. We make sure that it is
safe to remove the muxer once the EOS event is retrieved from its
source pad.

Change-Id: I5ddeb63797e48f10d2c220792a5efd53deb3b2fc