MediaSet: Do not destroy mediaset by default
Destroying the mediaset can cause problems because static variables
destruction is not thread save. This can cause deadlocks, bad
memory usage race conditions and this kind of problems.

Not releasing this object is not a problem, because the program is
finishing and the operative system is going to clean this memory.

Change-Id: I01090a0e928b30497763752e5de7167233393693