kmssdpcontext: Do not mask internal errors.
If an error happens, we must notify it using GError parameter. In this case
and SDP message was not well generated and this was not being notifying to
the user of the library who thinks that the resulting SDP message is well
generated. This is not a problem with offers, but it might generate unmatched
medias in answers when an error happens because the media
is ommited.

Change-Id: I562543705f16cb1c1315fe74db331a6b4828c8d5