remb: make increment phase more aggressive
Now, the remb base took into account is the max of
the last REMB and the max bitrate.
With this, we avoid stopping raising the REMB value
if the sender peer does not increase the bitrate for some
reason.
Anyway, we limit the REMB value to the double of the max bitrate.

Change-Id: I681315631d3176c4007c6dbc27a1daa1d2d2fed9