RtpSync tests: Generalize test 'test_sync_two_streams'
This test, as most others in this test suite, was working on the
assumption that initial (reference) time was zero.
In order to make the test more useful and to clarify the different
concepts about what is the reference time, the real-world NTP clock, and
the stream timestamps, each value has been separated into different
arbitrary values.