Rename kms_utils_caps_are_* to kms_utils_caps_is_*
Reasoning:
1. Convention: '_is_' is the de-facto name for functions that query some
property and return a boolean value. Having '_are_' functions is totally
unconventional and counter-intuitive for readers of the code.
2. Semantics: Even if "caps" mean "capabilities", which is a plural
word, and GstCaps objects contain (in theory, not so much in practice) a
list of structs, the fact is that '_is_ functions are called over *ONE*
thing at a time. It would make sense to use '_are_ for a function that
took an *array* of things, but it breaks the expected use of the name in
case of functions that take *one* single thing, regardless of what is
the name of that thing.