HubPort: Don't unlink all SRC pads unconditionally
Audio and video pads are always linked internally by classes that derive
from BaseHub, so it was sensible to assume that a pad would always exist
when unlinking.
But now that data pads have been added to BaseHub, not all derived
classes have been updated to handle those (currently only
compositemixer), so data pads are considered optional in this code.