KmsAgnosticBin: FIXME Reconfigure pipeline (Kurento/bugtracker#144)
Work on Kurento/bugtracker#144

When an RTSP stream is dosconnected and re-connected again, and the
PlayerEndpoint is stopped and later restarted, there is an internal
error in the pad chain function, somewhere between the Agnosticbin's
input fakesing and the parsetreebin.

This fix will replace the fakesink with a new one whenever a flow error
happens, and will also force reconfiguring the treebin.

It's been tested that these two operation must be done together. It's
not enough to either just replace the fakesink, or just reconfigure the
input bin.