Update kmsremb.c
Remove unwanted RTCP packet. When no REMB packet is sent (stats not updated) the RTCP packet previously added to the buffer in line 478 is not removed. That results on an empty RTCP packet added at the end of the UDP packet, with PT ==206 (REMB) but completely empty. Its impact should be minor, as most receivers should just discard that empty RTCP packet, but anyway is garbage in the stream