sdpagent: fix GStreamer-CRITICAL caused by multiple m= lines
Openvidu-Call is generating a 2-video and 2-audio Unified-Plan style SDP
Offer when running from desktop Safari.

The causes for the multi-track SDP are unknown (probably a bug in the JS
client code of openvidu-browser), but what we know is that KMS should
not try to configure twice the same RtpBin, once for each m= line.

The code to detect repeated m= lines is there, but it wasn't being
checked in order to avoid triggering the transport connection callback.