MediaElement: fix 100% CPU: avoid searching for pad name
MediaElementImpl::disconnect(), called from
MediaElementImpl::disconnectAll(), had to perform multiple searchs in
order to find the desired GstPad, only to then grab its name and pass it
to the C plugin, which would then have to search again for that name in
the list of element pads.

Iterating over all the connected pads multiple times, first on the C++
side and later on the C side, is very inefficient, and in our tests with
8000 connected endpoints, the CPU was wasted in mere lookups trying to
find a GstPad by its name. Depending on the server it could take up to
10 minutes with the CPU at 100%!

Instead, let's try this: upon creation we already have a pointer to the
GstPad. Why store its name and later have to search for it multiple
times? It is probably more efficient to just store the pointer and use
it directly whenever the element wants to disconnect (see private class
ElementConnectionDataInternal)

Fixes https://github.com/Kurento/bugtracker/issues/462
Fixes https://github.com/Kurento/bugtracker/issues/468