recording: Fix wrong caps for AAC audio
The MKV profile uses AAC audio, which the AAC encoders in GStreamer
(checked with voaacenc and avenc_aac) identify with the caps
"audio/mpeg,mpegversion=4", not "audio/x-aac".

This caused the agnosticbin to fail finding an acceptable encoder to
perform the audio transcoding, and the recording would end up not
working with an infinite stream of ERROR messages.

Fixes https://github.com/Kurento/bugtracker/issues/469