uriendpoint (PlayerEndpoint): avoid losing Base64 encoding from URIs
The call to gst_uri_get_location() would decode Base64-encoded URIs.
This was an unintended consequence because at this point in code we're
just parsing the string and still want the encoded version.

For example, an URI like:
rtsp://admin:passw%40rd@example.com

would be parsed as:

location = admin:passw@rd@example.com

and this would cause problems later when GStreamer tries to parse the
password and the hostname part of the URI.

Fixes https://github.com/Kurento/bugtracker/issues/554