rtp: use G_MAXUINT as invalid SSRC for initialization
The SSRC value 0 was used for initializing variables, and for error
checking, as if 0 was an invalid value. However, it is not. Neither is
G_MAXUINT, by the way: the whole range of unsigned 32-bit values are
valid SSRC. However we need to use _something_ for "invalid", and to be
honest the ideal solution would be a Maybe type. So, without changing
the existing code too much, using G_MAXUINT seems better than 0.

Mediasoup uses 0 as a valid SSRC, and sends that as its Source SSRC in
RTCP Receiver Reports.