rtp: remove rtcpdemux from "rtcp-mux" connections
With RTCP-mux connections (those that send both RTP and RTCP through the
same UDP port, with a different port for each single media; as
negotiated in SDP with "a=rtcp-mux"), the upstream element is already
providing RTP and RTCP in separate source pads.

This means we don't really have any need to use rtcpdemux or any other
element to demux buffers and separate RTP from RTCP: they already come
separated from upstream!

So the use of rtcpdemux in here didn't make any sense to begin with.