kmd: Remove dependency on kurento-client for generated pom.xml
Having code.api.java.mavenArtifactId be "kurento-client" introduced a
chicken-and-egg problem, because:

* Extra Kurento modules (i.e. any which is not core, elements, or
filters, such as kms-platedetector) would have a generated pom.xml that
depends on kurento-client.

* kurento-client depends on extra Kurento modules.

So, it wouldn't be possible to release a completely new version (i.e.
6.18.0) where ALL projects have been updated! Extra modules would depend
on the new version of kurento-client, but the new kurento-client would
depend on the new version of the extra modules. Ouch!

So, instead of declaring a dependency with "kurento-client" here, change
it to the self-generated "kms-api-*" one that is standalone and this
package is able to generate by itself, thus solving the issue.