bugfix for framerate detection in vp8parse
We have detected cases where the algorithm used in kmsvp8parse.c incorrectly generates abnormally huge framerates (even 1000/1. In Kurento this have a serious impact in memory as elements like videorate are forced to generate that framerate which is for sure incorrect. Most probably it may be generated when some PTS is estimated or just added 1 ms from previous frame. But even more, vp8 in wbertc is of variable framerate by nature. So it seems more correct to ensure that vp8parse always generates a variable framerate instead of trying to guess that value, that has shown to be a risk for the server stability