bugfix for framerate detection in vp8parse (#29)
## What is the current behavior you want to change?
We have detected cases where the algorithm used in kmsvp8parse.c incorrectly generates abnormally huge framerates (even 1000/1. In Kurento this have a serious impact in memory as elements like videorate are forced to generate that framerate which is for sure incorrect. Most probably it may be generated when some PTS is estimated or just added 1 ms from previous frame. But even more, vp8 in wbertc is of variable framerate by nature. So it seems more correct to ensure that vp8parse always generates a variable framerate instead of trying to guess that value, that has shown to be a risk for the server stability

## What is the new behavior provided by this change?
It just makes vp8parse always report a variable framerate regardless of what video is being ingested.

## How has this been tested?
on Kurento sample applications forcing the use of vp8 as codec an in a test application that uses a modified composite element. As expected, behaviour looked as before the change, but when we dump a .dot file we can assess that the output of vp8parse states a framerate of 0/1