MediaElement: Add "TargetEncoderBitrate" property + Min/Max family
The old "OutputBitrate" properties, introduced in
https://github.com/Kurento/kms-core/commit/f2392542e86d511efc2c0f0bc92aaabcef4ad890
and
https://github.com/Kurento/kms-core/commit/8442de9d6a3116a4425bc9bb559e9c4b7fc6a7eb,
had several problems:

* They were confusing. The naming of these properties is only vaguely
related to their exact behavior, and the "Output" word only added
confusion given all the other parameters that exist, such as all the
combinations of "(Max|Min)(Audio|Video)(Recv|Send)Bandwidth". In
reality, the "Output" properties had nothing to do with those, as they
pertained exclusively to the _output bitrate of the  video encoder_,
then transcoding is activated within the media server.

* They didn't do what they promised. Only Min and Max properties where
exposed, but the actual target bitrate was still hardcoded at 300 kbps.
This means that it wouldn't be possible to have a minimum below of the
desired target bitrate. For example, it was impossible to have a minimum
of 100 kbps, an encoding bitrate of 500 kbps, and a maximum of 1 Mbps.
This caused even more confusion, because users didn't understand why
their transcoded recordings where getting only 300 kbps even though they
wanted a higher quality output.

With this change, it it very clear to both users and developers what
there values are about. Also, the addition of "TargetEncoderBitrate"
allows to request a bitrate different from the previously hardcoded 300
kbps, even without touching the Min and Max limits.

Docs have also been extended to explain what these values do.